refactor: CartRepository 변경에 따른 코드 개선

+++ app/src/main/java/woowacourse/shopping/data/InMemoryCartRepository.kt
@@ -13,6 +13,6 @@ class InMemoryCartRepository : CartRepository {
     override suspend fun getAllCartProducts(): List<Product> = cartProducts.toList()
 
     override suspend fun deleteCartProduct(id: Long) {
-        cartProducts.removeAt(id)
+        cartProducts.removeAt(id.toInt())
     }
 }
